pyproject.toml
[project]
name = "model_registry"
version = "0.1.0"
description = "A public registry of LLM models from various providers."
requires-python = ">=3.11"
dependencies = [
    "pydantic",
    "requests",
    "python-dotenv"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "responses"
]

[build-system]
requires = ["uv", "uv_build"]
build-backend = "uv_build" 